fix: client-side server-side data fetching
