+++ src/components/supporters/contributors.tsx
@@ -1,8 +1,7 @@
 import RenderLogos from "../reusables/render-logo";
-import { fetchLogo } from "~/app/actions/supporters";
+import { fetchLogo } from "~/lib/supporters";
 
 const Members = async () => {
-
   const logos: Logo[] = await fetchLogo();
 
   return (
@@ -13,8 +12,7 @@ const Members = async () => {
         </div>
         <div className="grid grid-cols-1 md:grid-cols-3 items-center justify-center lg:grid-cols-3 gap-3 pb-8 md:pb-3 px-4 md:px-4">
           {RenderLogos(
-            logos
-              .sort((a, b) => a.name.localeCompare(b.name)),
+            logos.sort((a, b) => a.name.localeCompare(b.name)),
             "lg:self-center lg:flex lg:justify-center pb-2"
           )}
         </div>
